Update README.md
I am delighted to share about a project I worked on that aimed to improve patient care in Sri Lanka's healthcare sector. Our team developed an automatic IV fluid flow regulator to aid in the treatment of dengue patients, who require constant fluid intake. With the increasing number of patients during the rainy season, it is challenging for healthcare workers to treat patients manually, especially during the night shift. The device we developed enables patients to receive a constant supply of fluid during their sleep, eliminating the need for manual administration by nurses every hour. To develop this device, we conducted extensive research and designed a system that takes in the required hourly fluid volume and the number of hours the supply is needed. We encountered several challenges during the design process, but our team worked tirelessly to overcome them. I am proud to say that we successfully created an innovative solution that not only improves patient care but also alleviates the workload of healthcare workers. This project taught us a lot and allowed us to contribute towards addressing a significant issue in Sri Lanka's healthcare sector.